feat: 리뷰 시스템 구현 및 s3 파일 업로드 구현

+++ viewlog/src/main/java/com/trip/viewlog/attraction/infrastructure/jpaadapter/AttractionJpaRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface AttractionJpaRepository extends JpaRepository<AttractionEntity, Integer> {
+public interface AttractionJpaRepository extends JpaRepository<AttractionEntity, Long> {
 
 	@EntityGraph(attributePaths = "contentTypeEntity")
 	List<AttractionEntity> findByLatitudeBetweenAndLongitudeBetween(double minLat, double maxLat, double minLng, double maxLng);
